Renamed 'craco' to 'MC_sample'
