+++ zdm/MC_sample/testing.py
@@ -17,7 +17,7 @@ from matplotlib import pyplot as plt
 import pandas
 
 from zdm import iteration as it
-from zdm.craco import loading
+from zdm.MC_sample import loading
 
 from IPython import embed
